+++ src/main/java/dinkplugin/notifiers/KillCountNotifier.java
@@ -8,6 +8,7 @@ import dinkplugin.notifiers.data.BossNotificationData;
 import dinkplugin.util.KillCountService;
 import dinkplugin.util.TimeUtils;
 import dinkplugin.util.Utils;
+import lombok.Value;
 import lombok.extern.slf4j.Slf4j;
 import net.runelite.api.Client;
 import net.runelite.api.Varbits;
@@ -20,6 +21,7 @@ import org.apache.commons.lang3.tuple.Pair;
 import org.jetbrains.annotations.Nullable;
 import org.jetbrains.annotations.VisibleForTesting;
 
+import javax.inject.Inject;
 import javax.inject.Singleton;
 import java.time.Duration;
 import java.util.Optional;
@@ -40,7 +42,7 @@ public class KillCountNotifier extends BaseNotifier {
 
     private static final Pattern PRIMARY_REGEX = Pattern.compile("Your (?<key>.+)\\s(?<type>kill|chest|completion|harvest)\\s?count is: ?(?<value>[\\d,]+)\\b", Pattern.CASE_INSENSITIVE);
     private static final Pattern SECONDARY_REGEX = Pattern.compile("Your (?:completed|subdued) (?<key>.+) count is: (?<value>[\\d,]+)\\b");
-    private static final Pattern TIME_REGEX = Pattern.compile("(?:Duration|time|Subdued in):? (?<time>[\\d:]+(.\\d+)?)\\.?", Pattern.CASE_INSENSITIVE);
+    private static final Pattern TIME_REGEX = Pattern.compile("(?:Duration|time|Subdued in):? (?<time>[\\d:]+(?:.\\d+)?)\\.?(?: Personal best: (?<pbtime>[\\d:+]+(?:.\\d+)?))?", Pattern.CASE_INSENSITIVE);
 
     private static final String BA_BOSS_NAME = "Penance Queen";
 
@@ -52,6 +54,9 @@ public class KillCountNotifier extends BaseNotifier {
      */
     @VisibleForTesting
     static final int MAX_BAD_TICKS = 10;
+
+    @Inject
+    private KillCountService kcService;
 
     private final AtomicInteger badTicks = new AtomicInteger();
     private final AtomicReference<BossNotificationData> data = new AtomicReference<>();
@@ -92,7 +97,7 @@ public class KillCountNotifier extends BaseNotifier {
             // https://oldschool.runescape.wiki/w/Barbarian_Assault/Rewards#Earning_Honour_points
             if (widget != null && widget.getText().contains("80 ") && widget.getText().contains("5 ")) {
                 int gambleCount = client.getVarbitValue(Varbits.BA_GC);
-                this.data.set(new BossNotificationData(BA_BOSS_NAME, gambleCount, "The Queen is dead!", null, null, null));
+                this.data.set(new BossNotificationData(BA_BOSS_NAME, gambleCount, "The Queen is dead!", null, null, null, null));
             }
         }
     }
@@ -121,12 +126,20 @@ public class KillCountNotifier extends BaseNotifier {
             return;
 
         // ensure interval met or pb or ba, depending on config
+        boolean isPb = data.isPersonalBest() == Boolean.TRUE;
         boolean ba = data.getBoss().equals(BA_BOSS_NAME);
-        if (!checkKillInterval(data.getCount(), data.isPersonalBest()) && !ba)
+        if (!checkKillInterval(data.getCount(), isPb) && !ba)
             return;
 
+        // populate personalBest if absent
+        if (data.getPersonalBest() == null && !isPb) {
+            Duration pb = kcService.getPb(data.getBoss());
+            if (pb != null && (data.getTime() == null || pb.compareTo(data.getTime()) < 0)) {
+                data = data.withPersonalBest(pb);
+            }
+        }
+
         // Assemble content
-        boolean isPb = data.isPersonalBest() == Boolean.TRUE;
         String player = Utils.getPlayerName(client);
         String time = TimeUtils.format(data.getTime(), TimeUtils.isPreciseTiming(client));
         Template content = Template.builder()
@@ -147,8 +160,8 @@ public class KillCountNotifier extends BaseNotifier {
             .build());
     }
 
-    private boolean checkKillInterval(int killCount, @Nullable Boolean pb) {
-        if (pb == Boolean.TRUE && config.killCountNotifyBestTime())
+    private boolean checkKillInterval(int killCount, boolean pb) {
+        if (pb && config.killCountNotifyBestTime())
             return true;
 
         if (killCount == 1 && config.killCountNotifyInitial())
@@ -174,6 +187,7 @@ public class KillCountNotifier extends BaseNotifier {
                     defaultIfNull(updated.getGameMessage(), old.getGameMessage()),
                     updated.getTime() == null || (tob && old.getTime() != null) ? old.getTime() : updated.getTime(),
                     updated.isPersonalBest() == null || (tob && old.isPersonalBest() != null) ? old.isPersonalBest() : updated.isPersonalBest(),
+                    updated.getPersonalBest() == null || (tob && old.getPersonalBest() != null) ? old.getPersonalBest() : updated.getPersonalBest(),
                     defaultIfNull(updated.getParty(), old.getParty())
                 );
             }
@@ -184,21 +198,23 @@ public class KillCountNotifier extends BaseNotifier {
         if (message.startsWith("Preparation")) return Optional.empty();
         Optional<Pair<String, Integer>> boss = parseBoss(message);
         if (boss.isPresent())
-            return boss.map(pair -> new BossNotificationData(pair.getLeft(), pair.getRight(), message, null, null, Utils.getBossParty(client, pair.getLeft())));
+            return boss.map(pair -> new BossNotificationData(pair.getLeft(), pair.getRight(), message, null, null, null, Utils.getBossParty(client, pair.getLeft())));
 
         // TOB reports final wave duration before challenge time in the same message; skip to the part we care about
         int tobIndex = message.startsWith("Wave") ? message.indexOf(KillCountService.TOB) : -1;
         String msg = tobIndex < 0 ? message : message.substring(tobIndex);
 
-        return parseTime(msg).map(t -> new BossNotificationData(tobIndex < 0 ? null : KillCountService.TOB, null, null, t.getLeft(), t.getRight(), null));
+        return parseTime(msg).map(t -> new BossNotificationData(tobIndex < 0 ? null : KillCountService.TOB, null, null, t.getTime(), t.isPb(), t.getPb(), null));
     }
 
-    private static Optional<Pair<Duration, Boolean>> parseTime(String message) {
+    private static Optional<ParsedTime> parseTime(String message) {
         Matcher matcher = TIME_REGEX.matcher(message);
         if (matcher.find()) {
             Duration duration = TimeUtils.parseTime(matcher.group("time"));
-            boolean pb = message.toLowerCase().contains("(new personal best)");
-            return Optional.of(Pair.of(duration, pb));
+            boolean isPb = message.toLowerCase().contains("(new personal best)");
+            String pbTime = matcher.group("pbtime");
+            Duration pb = pbTime != null ? TimeUtils.parseTime(pbTime) : null;
+            return Optional.of(new ParsedTime(duration, isPb, pb));
         }
         return Optional.empty();
     }
@@ -270,5 +286,12 @@ public class KillCountNotifier extends BaseNotifier {
             return boss;
 
         return null;
+    }
+
+    @Value
+    private static class ParsedTime {
+        Duration time;
+        boolean isPb;
+        @Nullable Duration pb;
     }
 }