+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Minor: Add `personalBest` time to killcount notification metadata. (#648)
 - Minor: Fire loot notifications for new Royal Titans bosses. (#650)
 - Minor: Include Bran pet name in pet notifications. (#649)
 - Minor: Add `%SENDER%` template variable for chat notifier. (#644)